requirements.txt
Flask==2.3.3
pyjokes==0.6.0
python-dotenv==1.0.0
deep-translator==1.11.4
requests==2.31.0